Changed the type of the indices

+++ include/timed_condition_set.hh
@@ -64,7 +64,7 @@ namespace learnta {
         for (auto it2 = std::next(it); it2 != timedConditionsWithSize.end(); it2++) {
           // Check if the convex hull is the exact union
           auto convexHull = timedCondition.convexHull(it2->first);
-          if (convexHull.enumerate().size() == it->second + it2->second) {
+          if (convexHull.enumerate().size() == static_cast<std::size_t>(it->second + it2->second)) {
             it->first = std::move(convexHull);
             it->second += it2->second;
             timedConditionsWithSize.erase(it2);
